% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_name.R
\name{embedName}
\alias{embedName}
\title{embedName}
\value{
embedName object with methods for embedding names and calculating cosine similarity between names
}
\description{
Class for generating name embeddings
}
\examples{
embedder <- embedName$new(embedding_model = 'SamFrederick/namespace1m',)
embedder$embed_name('John Smith')

## ------------------------------------------------
## Method `embedName$embed_name`
## ------------------------------------------------

embedder <- embedName$new(embedding_model = 'SamFrederick/namespace1m',)
embedder$embed_name('John Smith')

## ------------------------------------------------
## Method `embedName$cosine_similarity`
## ------------------------------------------------

embedder <- embedName$new(embedding_model = 'SamFrederick/namespace1m',)
embedder$cosine_similarity('John Smith', 'Alex Johnson')
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-embedName-new}{\code{embedName$new()}}
\item \href{#method-embedName-embed_name}{\code{embedName$embed_name()}}
\item \href{#method-embedName-cosine_similarity}{\code{embedName$cosine_similarity()}}
\item \href{#method-embedName-clone}{\code{embedName$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-embedName-new"></a>}}
\if{latex}{\out{\hypertarget{method-embedName-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{embedName$new(
  embedding_model = "SamFrederick/namespace1m",
  device = "cpu",
  tokenizer = "roberta-large"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{embedding_model}}{string, default "SamFrederick/namematch1m". Embedding model to use for embedding names}

\item{\code{device}}{string, default "cpu". Can be "cpu", "cuda", "mps"}

\item{\code{tokenizer}}{string, default "roberta-large". A tokenizer with which to tokenize names for the embedding_model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-embedName-embed_name"></a>}}
\if{latex}{\out{\hypertarget{method-embedName-embed_name}{}}}
\subsection{Method \code{embed_name()}}{
A method to generate embeddings for name(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{embedName$embed_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a string or list of strings containing names for which to generate embeddings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a torch tensor containing the name embeddings
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{embedder <- embedName$new(embedding_model = 'SamFrederick/namespace1m',)
embedder$embed_name('John Smith')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-embedName-cosine_similarity"></a>}}
\if{latex}{\out{\hypertarget{method-embedName-cosine_similarity}{}}}
\subsection{Method \code{cosine_similarity()}}{
A method to calculate the cosine similarity between names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{embedName$cosine_similarity(name1, name2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name1, name2}}{a string or list of strings containing names for which to calculate cosine similarity}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a torch tensor containing the cosine similarities between names
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{embedder <- embedName$new(embedding_model = 'SamFrederick/namespace1m',)
embedder$cosine_similarity('John Smith', 'Alex Johnson')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-embedName-clone"></a>}}
\if{latex}{\out{\hypertarget{method-embedName-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{embedName$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
